rt version of distcorr warps
